Handle invalid URLs gracefully
This patch adds an app.routing module that is the entry point for all other routings. It includes an urlRouterProvider.otherwise() that handles unmatched states, and a generic '404' state that is used to display an html template in the event of an invalid url.

Reviewed-by: aazores
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-May/023213.html