Update timestampToDateFilter to replace unixToDateFilter
This patch makes updates to timestampToDateFilter, that allows it to replace the functionality of the former unixToDateFilter. This patch removes the dependency on 'moment', and instead uses the Angular date filter for converting epoch time to time strings. Additionally, a constant called DATE_FORMAT is included in the filters module that can be used to specify how the time string should be formatted.

Reviewed-by: aazores
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-July/023994.html