Use pfListView & pfPagination for jvm-list
This patch introduces the PatternFly listview (& toolbar) and pagination directives to jvm-list. Additionally, there are changes to the jvms mock-endpoint so the toolbar actions (sorting & searching) can be tested in a development setting. The dependency for angular-patternfly also had to be updated from 3.23.3 to 4.4.1.

Reviewed-by: aazores
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-August/024633.html