Determine web-gateway URL at runtime
Web-Client now expects the deploying webserver to respond to GET /gatewayurl
with a JSON response pointing the client to its associated Web-Gateway
instance. This allows the Web-Client to be built once and distributed for
deployment, rather than requiring a rebuild with the GATEWAY_URL environment
variable set. If /gatewayUrl is not implemented then the client makes a best-
effort guess that the deploying webserver is the Web-Gateway.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-September/025022.html